pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-gst-compliance-gym"
version = "0.1.0"
description = "OpenEnv environment for Indian GST compliance auditing"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.2",
    "fastapi>=0.115.0",
    "pydantic>=2.0.0",
    "uvicorn>=0.24.0",
    "requests>=2.31.0",
    "fastmcp>=0.1.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0.0", "pytest-cov>=4.0.0"]
inference = ["openai>=1.0.0"]

[project.scripts]
server = "gst_compliance_gym.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["gst_compliance_gym", "gst_compliance_gym.server"]
package-dir = {"gst_compliance_gym" = ".", "gst_compliance_gym.server" = "server"}
